feat: admin

+++ backend/src/main/java/com/lendy/backend/cart/dto/CartItemDto.java
@@ -1,15 +1,13 @@
 package com.lendy.backend.cart.dto;
 
-public record CartItemDto (
-        Integer productOptionId,
-        Long productId,
-        String imageURL,
-        String productName,
-        String color,
-        String size,
-        Integer count,
-        Integer unitPrice,
-        Integer unitTotalPrice
-)
-{
+public record CartItemDto(
+                Integer productOptionId,
+                Integer productId,
+                String imageURL,
+                String productName,
+                String color,
+                String size,
+                Integer count,
+                Integer unitPrice,
+                Integer unitTotalPrice) {
 }
